moving footer, landingpage,navbar outside user folder
