+++ app/(dashboard)/layout.tsx
@@ -1,0 +1,26 @@
+import Landingpage from './landingPage'
+import Footer from './Footer'
+import Navbar from './Navbar'
+
+export const metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <html lang="en">
+      <body>
+      <Navbar/>
+      <Landingpage />
+      <Footer />
+      </body>
+     
+
+    </html>
+  )
+}
