+++ app/(dashboard)/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-function page() {
+const page = () => {
   return (
-    <div>pagasdasdse</div>
+    <div>Landing page </div>
   )
 }
