--- app/(dashboard)/user/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const user = () => {
-  return (
-    <div>user </div>
-  )
-}
-
-export default user
