+++ app/(dashboard)/user/sidenav/page.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+
 import {
   AiOutlineBell,
   AiOutlineCloudDownload,
@@ -17,7 +18,7 @@ import { FiActivity } from "react-icons/fi";
 
 type Props = {};
 
-function page({}: Props) {
+function sidenav({}: Props) {
   return (
     <div className="grid items-center justify-around h-full grid-flow-row grid-cols-1 gap-2 px-4 py-6 text-center bg-white w-72">
       <p className="text-2xl font-extrabold text-black">NAVIGATION</p>
@@ -77,4 +78,4 @@ function page({}: Props) {
   );
 }
 
-export default page;
+export default sidenav;
