+++ app/layout.tsx
@@ -1,18 +1,16 @@
 import './globals.css'
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
-      <body>{children}</body>
+    <html>
+      <head />
+      <body>
+        {children}
+        </body>
     </html>
   )
 }
