+++ package.json
@@ -18,6 +18,7 @@
     "next-auth": "^4.19.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-social-icons": "^5.15.0",
     "typescript": "4.9.5"
   },
   "devDependencies": {
